+++ tests/testthat/test_analysis_example_vanhouwelingen1993.r
@@ -28,9 +28,9 @@ test_that("results of the fixed-effects conditional logistic model are correct."
 
    ### compare with results on page 2275 (in text)
    expect_equivalent(coef(res), 0.1216, tolerance=.tol[["coef"]])
-   expect_equivalent(res$se, 0.0993, tolerance=.tol[["se"]])
-   expect_equivalent(res$ci.lb, -0.0730, tolerance=.tol[["ci"]])
-   expect_equivalent(res$ci.ub, 0.3162, tolerance=.tol[["ci"]]) ### 0.31 in paper (rounded a bit more heavily, so 32-bit and 64-bit versions give same result)
+   expect_equivalent(res$se, 0.0995, tolerance=.tol[["se"]])
+   expect_equivalent(res$ci.lb, -0.0734, tolerance=.tol[["ci"]])
+   expect_equivalent(res$ci.ub, 0.3165, tolerance=.tol[["ci"]]) ### 0.31 in paper (rounded a bit more heavily, so 32-bit and 64-bit versions give same result)
    expect_equivalent(c(logLik(res)), -53.6789, tolerance=.tol[["fit"]])
 
    ### run with control(dnchgcalc="dnoncenhypergeom")
@@ -52,22 +52,22 @@ test_that("results of the random-effects conditional logistic model are correct.
    expect_warning(res <- rma.glmm(measure="OR", ai=b.xci, n1i=nci, ci=b.xti, n2i=nti, data=dat, model="CM.EL", method="ML"))
 
    ### compare with results on page 2277 (in text)
-   expect_equivalent(coef(res), 0.1745, tolerance=.tol[["coef"]])
-   expect_equivalent(res$se, 0.1360, tolerance=.tol[["se"]])
-   expect_equivalent(res$ci.lb, -0.0919, tolerance=.tol[["ci"]])
-   expect_equivalent(res$ci.ub, 0.4410, tolerance=.tol[["ci"]])
-   expect_equivalent(c(logLik(res)), -52.9890, tolerance=.tol[["fit"]])
-   expect_equivalent(res$tau2, 0.1195, tolerance=.tol[["var"]])
+   expect_equivalent(coef(res), 0.1744, tolerance=.tol[["coef"]])
+   expect_equivalent(res$se, 0.1364, tolerance=.tol[["se"]])
+   expect_equivalent(res$ci.lb, -0.0929, tolerance=.tol[["ci"]])
+   expect_equivalent(res$ci.ub, 0.4417, tolerance=.tol[["ci"]])
+   expect_equivalent(c(logLik(res)), -52.9001, tolerance=.tol[["fit"]])
+   expect_equivalent(res$tau2, 0.1192, tolerance=.tol[["var"]])
 
    ### run with control(dnchgcalc="dnoncenhypergeom")
    expect_warning(res <- rma.glmm(measure="OR", ai=b.xci, n1i=nci, ci=b.xti, n2i=nti, data=dat, model="CM.EL", method="ML", control=list(dnchgcalc="dnoncenhypergeom")))
 
    ### no discrepancies
-   expect_equivalent(coef(res), 0.1745, tolerance=.tol[["coef"]])
-   expect_equivalent(res$se, 0.1360, tolerance=.tol[["se"]])
-   expect_equivalent(res$ci.lb, -0.0921, tolerance=.tol[["ci"]])
-   expect_equivalent(res$ci.ub, 0.4411, tolerance=.tol[["ci"]])
-   expect_equivalent(c(logLik(res)), -52.9890, tolerance=.tol[["fit"]])
-   expect_equivalent(res$tau2, 0.1195, tolerance=.tol[["var"]])
+   expect_equivalent(coef(res), 0.1744, tolerance=.tol[["coef"]])
+   expect_equivalent(res$se, 0.1364, tolerance=.tol[["se"]])
+   expect_equivalent(res$ci.lb, -0.0930, tolerance=.tol[["ci"]])
+   expect_equivalent(res$ci.ub, 0.4418, tolerance=.tol[["ci"]])
+   expect_equivalent(c(logLik(res)), -52.9901, tolerance=.tol[["fit"]])
+   expect_equivalent(res$tau2, 0.1192, tolerance=.tol[["var"]])
 
 })